Move trigger input dialog to editor v2 components
