+++ internal-packages/workflow-designer-ui/src/editor/v2/components/run-button.tsx
@@ -14,9 +14,9 @@ import {
 import { defaultName, useWorkflowDesigner } from "@giselle-sdk/giselle/react";
 import { PlayIcon } from "lucide-react";
 import { type ButtonHTMLAttributes, useMemo, useState } from "react";
-import { TriggerInputDialog } from "../../../header/ui/trigger-input-dialog";
 import { useActController } from "../../../hooks/use-act-controller";
 import { NodeIcon } from "../../../icons/node";
+import { TriggerInputDialog } from "./trigger-input-dialog";
 
 function NodeSelectItem({
 	node,
