+++ internal-packages/workflow-designer-ui/src/editor/v2/components/trigger-input-dialog/helpers.ts
